Tests: workaround for freebsd 8 in disable_symlinks tests.
There are no O_SEARCH on FreeBSD yet, and we use O_EXEC|O_DIRECTORY
instead.  Unfortunately on FreeBSD 8 the O_DIRECTORY flag does nothing,
and as a result openat(O_EXEC) fails with EACCESS while trying to open
a file.  Workaround is to make the file executable, thus allowing it
to correctly fail with ENOTDIR on next call.