Tests: fixed possible test suite hang on exit.
In some cases (for example, call of global object method in a function), test
suite can hang during exit, because temporary directory holding a pid file is
removed prior to executing global test object desctructor.

The root cause is that the directory removal occurs in the END block (in the
File::Temp package) and the order of executing destructors and END blocks is
not specified in Perl.

The fix is to remove temporary directory in the test object destructor
explicitly and get rid of 'CLEANUP' flag in the tempdir() call.  The temporary
directory is removed by the means of rmtree() function, called inside an eval
block to exclude the case when rmtree may abort (some versions are known
to do it).