Tests: fixed test failure in proxy_unfinished.t with big buffers.
Skip the "no proxy temp" test if the response is finished, namely, if a
calculated content length of the chunked body matches the expected value.

Although a finished response is sort of expected, it can be further split
in chunks with a recent change f379b32e4733 (1.11.2) in sub filter module,
which broke a pre-existing check for the finished response.