Tests: added exception test to stream_js.t using 'require'.
The stream js tests introduced in edf5a3c9e36a fail on njs 0.1.14.  It doesn't
currently provide an easy way to check its version, whilst we are obligated to
gracefully handle such cases somehow.  With such an addition of 'require', now
the tests are skipped instead on the previous versions.