Tests: testing worker_shutdown_timeout works with HTTP/2.
A request in HTTP/2 is ensured to be finalized with worker_shutdown_timeout
by testing that it is logged at least, and the status code is not set as it
would on proxy_read_timeout.  After the fix has been fixed in ab5cac9d3f00,
the latter can be tuned to a safer value without losing in execution time.