Tests: reworked ssl_verify_depth tests.
Full matrix of various verify depths and chains is now tested.
Incompatible behaviour of OpenSSL 1.1.0+, which now limits the total
length of a chain instead of maximum number of signatures checked,
is explained in the comments.  Attempts to incorrectly use
client-provided intermediate certificates, introduced in aa5a61d1254b,
are removed.