Tests: proxy_cache_use_stale.t speedup.
Tests for not blocked stale responses are adjusted to have a lesser new
response size.  It should be enough to delay sending a new response for
just a couple of seconds to catch the old bug fixed in 1.13.1.  This
allows to speed up shutting down nginx that waits for those responses.