Tests: revised tests for listen port ranges.
Renumbered testing ports to get more chance to execute when run in parallel.

Relaxed condition to skip tests only when the port range is out of sequence.
Adjacent port numbers out of a specified range aren't crucial to skip tests:
if not in sequence, statistically this will be caught in subsequent runs.

Unsafe tests that use wildcard addresses are moved to a separate file.