Tests: handling of EWOULDBLOCK from sysread() with IO::Socket::SSL.
With IO::Socket::SSL, when select() reports that the socket is readable,
reading from it might still fail with EWOULDBLOCK, since no application
data is available in the socket.  In particular, this might happen with
TLSv1.3 when a session ticket is received after the handshake.  Fix is
to explicitly check for EWOULDBLOCK errors.