Tests: separate SSL session reuse tests.
Instead of being mixed with generic SSL tests, session reuse variants
are now tested in a separate file.

In the generic SSL tests only basic session reuse is now tested,
notably with session tickets enabled and a shared SSL session cache.
This should make it possible to reuse sessions in all cases (except
when it's not supported, such as with LibreSSL with TLSv1.3).

Note that session reuse with tickets implies that $ssl_session_id
is selected by the client and therefore is not available on the
initial connection.  Relevant test is modified to handle this.

Further, BoringSSL does not use legacy session ID with TLSv1.3 even
if it is sent by the client.  In contrast, OpenSSL always generates
an unique legacy session id, so it is available with TLSv1.3 even if
session resumption does not work (such as with old Net::SSLeay and
IO::Socket::SSL modules).