Tests: improved test for a stream initiating QUIC path migration.
Previously, the test mostly passed due to a push timer scheduled for an
unrelated delayed ACK, and the new path had enough data received to send
response back when the push timer fired.  Otherwise, the test could fail
due to unvalidated address.  Now it is fixed to perform path validation.