Tests: fixed ClientHello with resending Initial QUIC packets.
Previously it was rebuilt each time using distinct ClientHello.random
resulting in different CRYPTO payload.  As such, it led to TLS digest
hash and derived secrets mismatch when resending Initial packet.  Now
ClientHello is built once and reused when resending Initial packets.

Additionally, this required to preserve a generated secret value used
in shared secret calculation as part of TLS key schedule.  Previously
it was regenerated when receiving a Retry packet, but this won't work
with reused ClientHello as the resulting shared secrets won't match.