Tests: adjusted session reuse tests in ssl_ocsp.t.
Similar to generic SSL tests, only basic session reuse is now tested,
notably with session tickets enabled and a shared SSL session cache.
This should make it possible to reuse sessions in all cases (except
when it's not supported, such as with LibreSSL with TLSv1.3).

Notably, this allows to pass tests with various OpenSSL forks, which
may provide OCSP support but lack TLSv1.3 session reuse with a shared
SSL session cache.