Tests: simplified http_server_name.t by using string compare.
Repetitive http header field names are now hidden in test subroutines, and
expecting field values are compared as a string, instead of using a regex.

This change is anticipated to inherit http_server_name.t in testing server
names in the stream module, and to make the tests look alike.