Tests: adjusted ssl_engine_keys.t to run on Linux and macOS.
Previously, library paths were hardcoded to FreeBSD.

Further, a new TEST_NGINX_SOFTHSM environment variable is used to
append the softhsm search path as needed.  Also, the "dynamic_path"
command appears to be required on FreeBSD only, where OpenSSL looks
for the pkcs11 engine in a path different from where it is installed.

Based on a patch by Oksana Deeva.