Use CRLF where that is required
Lots of code used a bare LF in the HTTP protocol, but that is wrong:
RFC9112 is very clear that CRLF is the HTTP line terminator.

Most of the changes are just replacing \n with \r\n or adding missing \r
in heredocs, but it is also necessary to update byte counts to include
the extra bytes.