Tests: skip ssl_store_keys.t on win32.
The OSSL_STORE API normally expects that "uri" with the "file" scheme
is provided as an absolute path represented starting either with "/"
in the "path-absolute" rule, or, specifically for win32, with a drive
letter prefixed with "/" (see RFC 8089 for details).

Since both forms cannot be represented simultaneously without excessive
complexity introduced, skip tests for now on win32.