Tests: simplified QUIC tests for discarding 0-RTT keys.
Testing for regressions in discarding 0-RTT keys is now separated
from 1-RTT stream tests.  Completing a QUIC handshake in tests
includes sending some 1-RTT data in control and encoder streams.
This is enough to trigger the bug fixed in 930caed3b (1.27.4).

Also, this improves test stability due to a pending bug in HTTP3.pm,
which causes certain QUIC frames to be not reported when received
together with stream frames in a single packet.